+++ api/builtins_qlik/FullPath_test.go
@@ -1,4 +1,4 @@
-package main
+package builtins_qlik
 
 import (
 	"fmt"
@@ -6,11 +6,12 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"sigs.k8s.io/kustomize/api/builtins_qlik/utils/loadertest"
+	"sigs.k8s.io/kustomize/api/internal/k8sdeps/transformer"
 	"sigs.k8s.io/kustomize/api/k8sdeps/kunstruct"
 	"sigs.k8s.io/kustomize/api/resmap"
 	"sigs.k8s.io/kustomize/api/resource"
 	valtest_test "sigs.k8s.io/kustomize/api/testutils/valtest"
-	"sigs.k8s.io/kustomize/plugin/builtins_qlik/utils/loadertest"
 )
 
 func TestFullPath(t *testing.T) {
@@ -147,19 +148,21 @@ fieldSpecs:
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
 			resourceFactory := resmap.NewFactory(resource.NewFactory(
-				kunstruct.NewKunstructuredFactoryImpl()), nil)
+				kunstruct.NewKunstructuredFactoryImpl()), transformer.NewFactoryImpl())
 
 			resMap, err := resourceFactory.NewResMapFromBytes([]byte(testCase.pluginInputResources))
 			if err != nil {
 				t.Fatalf("Err: %v", err)
 			}
 
-			err = KustomizePlugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader(testCase.loaderRootDir), valtest_test.MakeFakeValidator(), resourceFactory), []byte(testCase.pluginConfig))
+			plugin := NewFullPathPlugin()
+
+			err = plugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader(testCase.loaderRootDir), valtest_test.MakeFakeValidator(), resourceFactory), []byte(testCase.pluginConfig))
 			if err != nil {
 				t.Fatalf("Err: %v", err)
 			}
 
-			err = KustomizePlugin.Transform(resMap)
+			err = plugin.Transform(resMap)
 			if err != nil {
 				t.Fatalf("Err: %v", err)
 			}
