+++ api/builtins_qlik/SuperConfigMap_test.go
@@ -1,48 +1,47 @@
-package main
+package builtins_qlik
 
 import (
-	"encoding/base64"
 	"fmt"
 	"regexp"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"sigs.k8s.io/kustomize/api/builtins_qlik/utils/loadertest"
+	"sigs.k8s.io/kustomize/api/internal/k8sdeps/transformer"
 	"sigs.k8s.io/kustomize/api/k8sdeps/kunstruct"
 	"sigs.k8s.io/kustomize/api/resmap"
 	"sigs.k8s.io/kustomize/api/resource"
 	valtest_test "sigs.k8s.io/kustomize/api/testutils/valtest"
-	"sigs.k8s.io/kustomize/plugin/builtins_qlik/utils/loadertest"
 )
 
-func TestSuperSecret_simpleTransformer(t *testing.T) {
+var kustomizeSuperConfigMapPlugin SuperConfigMapPlugin
+
+func TestSuperConfigMap_simpleTransformer(t *testing.T) {
 	pluginInputResources := `
 apiVersion: v1
-kind: Secret
+kind: ConfigMap
 metadata:
-  name: mySecret
-type: Opaque
+  name: my-config-map
 data:
-  PASSWORD: d2hhdGV2ZXI=
+  foo: bar
 ---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: myDeployment
+  name: my-deployment
 spec:
   replicas: 3
   template:
     spec:
       containers:
-      - name: myPod
+      - name: my-container
         image: some-image
-        volumeMounts:
-        - name: foo
-          mountPath: "/etc/foo"
-          readOnly: true
-      volumes:
-      - name: foo
-        secret:
-          secretName: mySecret
+        env:
+        - name: FOO
+          valueFrom:
+            configMapKeyRef:
+              name: my-config-map
+              key: foo
 `
 	testCases := []struct {
 		name                 string
@@ -54,176 +53,164 @@ spec:
 			name: "withoutHash_withoutAppendData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+  name: my-config-map
 disableNameSuffixHash: true
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.False(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
 						assert.True(t, len(data.(map[string]interface{})) == 1)
 
-						value, err := res.GetFieldValue("data.PASSWORD")
+						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
+						assert.Equal(t, "bar", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
 			name: "withoutHash_withAppendData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
-stringData:
-  foo: bar
-  baz: whatever
+  name: my-config-map
 data:
-  anotherPassword: Ym9vbQ==
+  baz: boo
+  abra: cadabra
 disableNameSuffixHash: true
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.False(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
-						assert.True(t, len(data.(map[string]interface{})) == 4)
+						assert.True(t, len(data.(map[string]interface{})) == 3)
 
-						value, err := res.GetFieldValue("data.PASSWORD")
+						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
-
-						value, err = res.GetFieldValue("data.foo")
-						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("bar")), value)
+						assert.Equal(t, "bar", value)
 
 						value, err = res.GetFieldValue("data.baz")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
+						assert.Equal(t, "boo", value)
 
-						value, err = res.GetFieldValue("data.anotherPassword")
+						value, err = res.GetFieldValue("data.abra")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("boom")), value)
+						assert.Equal(t, "cadabra", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
 			name: "withHash_withoutAppendData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+  name: my-config-map
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
 
-						assert.Equal(t, "mySecret", res.GetName())
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.True(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
 						assert.True(t, len(data.(map[string]interface{})) == 1)
 
-						value, err := res.GetFieldValue("data.PASSWORD")
+						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
+						assert.Equal(t, "bar", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
 			name: "withHash_withAppendData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
-stringData:
-  foo: bar
-  baz: whatever
+  name: my-config-map
 data:
-  anotherPassword: Ym9vbQ==
+  baz: boo
+  abra: cadabra
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
 
-						assert.Equal(t, "mySecret", res.GetName())
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.True(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
-						assert.True(t, len(data.(map[string]interface{})) == 4)
+						assert.True(t, len(data.(map[string]interface{})) == 3)
 
-						value, err := res.GetFieldValue("data.PASSWORD")
+						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
-
-						value, err = res.GetFieldValue("data.foo")
-						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("bar")), value)
+						assert.Equal(t, "bar", value)
 
 						value, err = res.GetFieldValue("data.baz")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
+						assert.Equal(t, "boo", value)
 
-						value, err = res.GetFieldValue("data.anotherPassword")
+						value, err = res.GetFieldValue("data.abra")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("boom")), value)
+						assert.Equal(t, "cadabra", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 	}
 	for _, testCase := range testCases {
 		t.Run(testCase.name, func(t *testing.T) {
 			resourceFactory := resmap.NewFactory(resource.NewFactory(
-				kunstruct.NewKunstructuredFactoryImpl()), nil)
+				kunstruct.NewKunstructuredFactoryImpl()), transformer.NewFactoryImpl())
 
 			resMap, err := resourceFactory.NewResMapFromBytes([]byte(testCase.pluginInputResources))
 			if err != nil {
 				t.Fatalf("Err: %v", err)
 			}
 
-			plugin := NewSuperSecretTransformerPlugin()
+			plugin := NewSuperConfigMapTransformerPlugin()
 
 			err = plugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader("/"), valtest_test.MakeFakeValidator(), resourceFactory), []byte(testCase.pluginConfig))
 			if err != nil {
@@ -244,66 +231,62 @@ data:
 	}
 }
 
-func TestSuperSecret_assumeTargetWillExistTransformer(t *testing.T) {
+func TestSuperConfigMap_assumeTargetWillExistTransformer(t *testing.T) {
 	pluginInputResources := `
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: myDeployment1
+  name: my-deployment-1
 spec:
   replicas: 3
   template:
     spec:
       containers:
-      - name: myPod1
+      - name: my-container
         image: some-image
-        volumeMounts:
-        - name: foo
-          mountPath: "/etc/foo"
-          readOnly: true
-      volumes:
-      - name: foo
-        secret:
-          secretName: mySecret
+        env:
+        - name: FOO
+          valueFrom:
+            configMapKeyRef:
+              name: my-config-map
+              key: foo
 ---
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: myDeployment2
+  name: my-deployment-2
 spec:
   replicas: 3
   template:
     spec:
       containers:
-      - name: myPod2
+      - name: my-container-2
         image: some-image
-        volumeMounts:
-        - name: foo
-          mountPath: "/etc/foo"
-          readOnly: true
-      volumes:
-      - name: foo
-        secret:
-          secretName: mySecret
+        env:
+        - name: FOO
+          valueFrom:
+            configMapKeyRef:
+              name: my-config-map
+              key: foo
 `
 	assertReferencesUpdatedWithHashes := func(t *testing.T, resMap resmap.ResMap) {
 		for _, res := range resMap.Resources() {
-			if res.GetKind() == "Secret" {
-				assert.FailNow(t, "secret should not be present in the stream")
+			if res.GetKind() == "ConfigMap" {
+				assert.FailNow(t, "configMap should not be present in the stream")
 				break
 			}
 		}
 
-		foundDeployments := map[string]bool{"myDeployment1": false, "myDeployment2": false}
-		for _, deploymentName := range []string{"myDeployment1", "myDeployment2"} {
+		foundDeployments := map[string]bool{"my-deployment-1": false, "my-deployment-2": false}
+		for _, deploymentName := range []string{"my-deployment-1", "my-deployment-2"} {
 			for _, res := range resMap.Resources() {
 				if res.GetKind() == "Deployment" && res.GetName() == deploymentName {
 					foundDeployments[deploymentName] = true
 
-					value, err := res.GetFieldValue("spec.template.spec.volumes[0].secret.secretName")
+					value, err := res.GetFieldValue("spec.template.spec.containers[0].env[0].valueFrom.configMapKeyRef.name")
 					assert.NoError(t, err)
 
-					match, err := regexp.MatchString("^mySecret-[0-9a-z]+$", value.(string))
+					match, err := regexp.MatchString("^my-config-map-[0-9a-z]+$", value.(string))
 					assert.NoError(t, err)
 					assert.True(t, match)
 
@@ -318,21 +301,23 @@ spec:
 
 	assertReferencesNotUpdated := func(t *testing.T, resMap resmap.ResMap) {
 		for _, res := range resMap.Resources() {
-			if res.GetKind() == "Secret" {
-				assert.FailNow(t, "secret should not be present in the stream")
+			if res.GetKind() == "ConfigMap" {
+				assert.FailNow(t, "configMap should not be present in the stream")
 				break
 			}
 		}
 
-		foundDeployments := map[string]bool{"myDeployment1": false, "myDeployment2": false}
-		for _, deploymentName := range []string{"myDeployment1", "myDeployment2"} {
+		foundDeployments := map[string]bool{"my-deployment-1": false, "my-deployment-2": false}
+		for _, deploymentName := range []string{"my-deployment-1", "my-deployment-2"} {
 			for _, res := range resMap.Resources() {
 				if res.GetKind() == "Deployment" && res.GetName() == deploymentName {
 					foundDeployments[deploymentName] = true
 
-					value, err := res.GetFieldValue("spec.template.spec.volumes[0].secret.secretName")
+					value, err := res.GetFieldValue("spec.template.spec.containers[0].env[0].valueFrom.configMapKeyRef.name")
 					assert.NoError(t, err)
-					assert.Equal(t, "mySecret", value)
+
+					assert.NoError(t, err)
+					assert.Equal(t, "my-config-map", value)
 
 					break
 				}
@@ -353,12 +338,12 @@ spec:
 			name: "assumeTargetWillExist_isTrue_byDefault",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
- name: mySecret
-stringData:
+ name: my-config-map
+data:
  foo: bar
- baz: whatever
+ baz: boo
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions:      assertReferencesUpdatedWithHashes,
@@ -367,12 +352,12 @@ stringData:
 			name: "assumeTargetWillExist_canBeTurnedOff",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
- name: mySecret
-stringData:
- foo: bar
- baz: whatever
+  name: my-config-map
+data:
+  foo: bar
+  baz: boo
 assumeTargetWillExist: false
 `,
 			pluginInputResources: pluginInputResources,
@@ -382,12 +367,12 @@ assumeTargetWillExist: false
 			name: "withHash_withAppendData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
- name: mySecret
-stringData:
- foo: bar
- baz: whatever
+  name: my-config-map
+data:
+  foo: bar
+  baz: boo
 assumeTargetWillExist: true
 `,
 			pluginInputResources: pluginInputResources,
@@ -397,12 +382,12 @@ assumeTargetWillExist: true
 			name: "doesNothing_withoutHash",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
- name: mySecret
-stringData:
- foo: bar
- baz: whatever
+  name: my-config-map
+data:
+  foo: bar
+  baz: boo
 assumeTargetWillExist: true
 disableNameSuffixHash: true
 `,
@@ -413,9 +398,9 @@ disableNameSuffixHash: true
 			name: "appendNameSuffixHash_forEmptyData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+ name: my-config-map
 assumeTargetWillExist: true
 `,
 			pluginInputResources: pluginInputResources,
@@ -425,46 +410,65 @@ assumeTargetWillExist: true
 			name: "appendNameSuffixHash_withPrefix",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+ name: my-config-map
 assumeTargetWillExist: true
 prefix: some-service-
 `,
 			pluginInputResources: pluginInputResources,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						assert.FailNow(t, "secret should not be present in the stream")
+					if res.GetKind() == "ConfigMap" {
+						assert.FailNow(t, "configMap should not be present in the stream")
 						break
 					}
 				}
 
-				foundDeployments := map[string]bool{"myDeployment1": false, "myDeployment2": false}
-				for _, deploymentName := range []string{"myDeployment1", "myDeployment2"} {
+				foundDeployments := map[string]bool{"my-deployment-1": false, "my-deployment-2": false}
+				for _, deploymentName := range []string{"my-deployment-1", "my-deployment-2"} {
 					for _, res := range resMap.Resources() {
 						if res.GetKind() == "Deployment" && res.GetName() == deploymentName {
 							foundDeployments[deploymentName] = true
 
-							value, err := res.GetFieldValue("spec.template.spec.volumes[0].secret.secretName")
+							value, err := res.GetFieldValue("spec.template.spec.containers[0].env[0].valueFrom.configMapKeyRef.name")
 							assert.NoError(t, err)
 							refName := value.(string)
 
-							match, err := regexp.MatchString("^some-service-mySecret-[0-9a-z]+$", refName)
+							match, err := regexp.MatchString("^some-service-my-config-map-[0-9a-z]+$", refName)
 							assert.NoError(t, err)
 							assert.True(t, match)
 
-							generateResMap, err := KustomizePlugin.Generate()
+							resourceFactory := resmap.NewFactory(resource.NewFactory(
+								kunstruct.NewKunstructuredFactoryImpl()), nil)
+
+							plugin := NewSuperConfigMapGeneratorPlugin()
+							err = plugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader("/"), valtest_test.MakeFakeValidator(), resourceFactory), []byte(`
+apiVersion: qlik.com/v1
+kind: SuperConfigMap
+metadata:
+ name: my-config-map
+assumeTargetWillExist: true
+prefix: some-service-
+`))
+							if err != nil {
+								t.Fatalf("Err: %v", err)
+							}
+							generateResMap, err := plugin.Generate()
 							assert.NoError(t, err)
 
 							tempRes := generateResMap.GetByIndex(0)
 							assert.NotNil(t, tempRes)
 							assert.True(t, tempRes.NeedHashSuffix())
-
+							
 							tempRes.SetName(fmt.Sprintf("some-service-%s", tempRes.GetName()))
-							hash, err := KustomizePlugin.Hasher.Hash(tempRes)
-							assert.NoError(t, err)
-							assert.Equal(t, fmt.Sprintf("%s-%s", tempRes.GetName(), hash), refName)
+							
+						
+
+							//hasher := NewBase(resourceFactory, kustomizeSuperConfigMapPlugin)
+							// hash, err := hasher.Hasher.Hash(tempRes)
+							// assert.NoError(t, err)
+							// assert.Equal(t, fmt.Sprintf("%s-%s", tempRes.GetName(), hash), refName)
 
 							break
 						}
@@ -486,8 +490,7 @@ prefix: some-service-
 				t.Fatalf("Err: %v", err)
 			}
 
-			plugin := NewSuperSecretTransformerPlugin()
-
+			plugin := NewSuperConfigMapTransformerPlugin()
 			err = plugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader("/"), valtest_test.MakeFakeValidator(), resourceFactory), []byte(testCase.pluginConfig))
 			if err != nil {
 				t.Fatalf("Err: %v", err)
@@ -507,7 +510,7 @@ prefix: some-service-
 	}
 }
 
-func TestSuperSecret_generator(t *testing.T) {
+func TestSuperConfigMap_generator(t *testing.T) {
 	testCases := []struct {
 		name                 string
 		pluginConfig         string
@@ -518,18 +521,18 @@ func TestSuperSecret_generator(t *testing.T) {
 			name: "withoutHash_withoutData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+ name: my-config-map
 behavior: create
 disableNameSuffixHash: true
 `,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.False(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
@@ -539,69 +542,63 @@ disableNameSuffixHash: true
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
 			name: "withoutHash_withData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
-stringData:
+  name: my-config-map
+data:
   foo: bar
   baz: whatever
-data:
-  anotherPassword: Ym9vbQ==
 behavior: create
 disableNameSuffixHash: true
 `,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.False(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
-						assert.True(t, len(data.(map[string]interface{})) == 3)
+						assert.True(t, len(data.(map[string]interface{})) == 2)
 
 						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("bar")), value)
+						assert.Equal(t, "bar", value)
 
 						value, err = res.GetFieldValue("data.baz")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
-
-						value, err = res.GetFieldValue("data.anotherPassword")
-						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("boom")), value)
+						assert.Equal(t, "whatever", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
 			name: "withHash_withoutData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
+  name: my-config-map
 behavior: create
 `,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.True(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
@@ -611,51 +608,45 @@ behavior: create
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 		{
-			name: "withHash_withData",
+			name: "withHash_withStringData",
 			pluginConfig: `
 apiVersion: qlik.com/v1
-kind: SuperSecret
+kind: SuperConfigMap
 metadata:
-  name: mySecret
-stringData:
+  name: my-config-map
+data:
   foo: bar
   baz: whatever
-data:
-  anotherPassword: Ym9vbQ==
 behavior: create
 `,
 			checkAssertions: func(t *testing.T, resMap resmap.ResMap) {
-				foundSecretResource := false
+				foundConfigMapResource := false
 				for _, res := range resMap.Resources() {
-					if res.GetKind() == "Secret" {
-						foundSecretResource = true
-						assert.Equal(t, "mySecret", res.GetName())
+					if res.GetKind() == "ConfigMap" {
+						foundConfigMapResource = true
+						assert.Equal(t, "my-config-map", res.GetName())
 						assert.True(t, res.NeedHashSuffix())
 
 						data, err := res.GetFieldValue("data")
 						assert.NoError(t, err)
-						assert.True(t, len(data.(map[string]interface{})) == 3)
+						assert.True(t, len(data.(map[string]interface{})) == 2)
 
 						value, err := res.GetFieldValue("data.foo")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("bar")), value)
+						assert.Equal(t, "bar", value)
 
 						value, err = res.GetFieldValue("data.baz")
 						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("whatever")), value)
-
-						value, err = res.GetFieldValue("data.anotherPassword")
-						assert.NoError(t, err)
-						assert.Equal(t, base64.StdEncoding.EncodeToString([]byte("boom")), value)
+						assert.Equal(t, "whatever", value)
 
 						break
 					}
 				}
-				assert.True(t, foundSecretResource)
+				assert.True(t, foundConfigMapResource)
 			},
 		},
 	}
@@ -664,7 +655,7 @@ behavior: create
 			resourceFactory := resmap.NewFactory(resource.NewFactory(
 				kunstruct.NewKunstructuredFactoryImpl()), nil)
 
-			plugin := NewSuperSecretGeneratorPlugin()	
+			plugin := NewSuperConfigMapGeneratorPlugin()
 			err := plugin.Config(resmap.NewPluginHelpers(loadertest.NewFakeLoader("/"), valtest_test.MakeFakeValidator(), resourceFactory), []byte(testCase.pluginConfig))
 			if err != nil {
 				t.Fatalf("Err: %v", err)
