+++ api/go.mod
@@ -7,7 +7,9 @@ require (
 	github.com/go-openapi/spec v0.19.4
 	github.com/golangci/golangci-lint v1.21.0
 	github.com/google/shlex v0.0.0-20191202100458-e7afc7fbc510
+	github.com/imdario/mergo v0.3.5
 	github.com/pkg/errors v0.8.1
+	github.com/stretchr/testify v1.4.0
 	golang.org/x/tools v0.0.0-20191010075000-0337d82405ff
 	gopkg.in/yaml.v2 v2.2.4
 	k8s.io/api v0.17.0
