--- plugin/builtins_qlik/gomplate/go.mod
@@ -1,10 +1,0 @@
-module sigs.k8s.io/kustomize/plugin/builtins_qlik/gomplate
-
-go 1.13
-
-require (
-	sigs.k8s.io/kustomize/api v0.3.1
-	sigs.k8s.io/yaml v1.1.0
-)
-
-replace sigs.k8s.io/kustomize/api => ../../../api
