--- plugin/builtins_qlik/utils/go.mod
@@ -1,9 +1,0 @@
-module sigs.k8s.io/kustomize/plugin/builtins_qlik/utils
-
-go 1.13
-
-require (
-	github.com/stretchr/testify v1.4.0
-	sigs.k8s.io/kustomize/api v0.3.1 // indirect
-	sigs.k8s.io/kustomize/v3 v3.3.1
-)
