--- plugin/builtins_qlik/valuesfile/go.mod
@@ -1,12 +1,0 @@
-module sigs.k8s.io/kustomize/plugin/builtins_qlik/valuesfile
-
-go 1.13
-
-require (
-	github.com/imdario/mergo v0.3.5
-	github.com/stretchr/testify v1.4.0
-	sigs.k8s.io/kustomize/api v0.3.0
-	sigs.k8s.io/yaml v1.1.0
-)
-
-replace sigs.k8s.io/kustomize/api => ../../../api
